opencv eigen recognizer!

+++ client/src/main/java/co/rxstack/ml/client/aws/converter/BoundingBoxConverter.java
@@ -3,7 +3,6 @@ package co.rxstack.ml.client.aws.converter;
 import java.util.function.Function;
 
 import co.rxstack.ml.common.model.FaceRectangle;
-
 import com.amazonaws.services.rekognition.model.BoundingBox;
 
 /**
@@ -12,7 +11,7 @@ import com.amazonaws.services.rekognition.model.BoundingBox;
 public class BoundingBoxConverter implements Function<BoundingBox, FaceRectangle> {
 	@Override
 	public FaceRectangle apply(BoundingBox boundingBox) {
-		return new FaceRectangle(boundingBox.getLeft(), boundingBox.getTop(), boundingBox.getWidth(),
-			boundingBox.getHeight());
+		return new FaceRectangle(boundingBox.getLeft().intValue(), boundingBox.getTop().intValue(),
+			boundingBox.getWidth().intValue(), boundingBox.getHeight().intValue());
 	}
 }
